+++ emulator/go.mod
@@ -3,8 +3,9 @@ module application-emulator
 go 1.20
 
 require (
+	github.com/iancoleman/strcase v0.3.0
 	golang.org/x/sys v0.9.0
-	google.golang.org/grpc v1.56.2
+	google.golang.org/grpc v1.57.0
 	google.golang.org/protobuf v1.31.0
 )
 
@@ -12,5 +13,5 @@ require (
 	github.com/golang/protobuf v1.5.3 // indirect
 	golang.org/x/net v0.9.0 // indirect
 	golang.org/x/text v0.9.0 // indirect
-	google.golang.org/genproto v0.0.0-20230410155749-daa745c078e1 // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20230525234030-28d5490b6b19 // indirect
 )
