+++ emulator/src/server/grpc.go
@@ -29,6 +29,8 @@ import (
 	"google.golang.org/grpc/health/grpc_health_v1"
 	"google.golang.org/grpc/reflection"
 	"google.golang.org/grpc/status"
+
+	"github.com/iancoleman/strcase"
 )
 
 type HealthServerImpl struct {
@@ -36,7 +38,7 @@ type HealthServerImpl struct {
 }
 
 func (h *HealthServerImpl) Check(ctx context.Context, request *grpc_health_v1.HealthCheckRequest) (*grpc_health_v1.HealthCheckResponse, error) {
-	grpcServiceName := fmt.Sprintf("generated.%s", model.GoName(util.ServiceName))
+	grpcServiceName := fmt.Sprintf("generated.%s", strcase.ToCamel(util.ServiceName))
 	if request.Service == "" || request.Service == grpcServiceName {
 		return &grpc_health_v1.HealthCheckResponse{
 			Status: grpc_health_v1.HealthCheckResponse_SERVING,
