+++ generator/go.mod
@@ -3,6 +3,7 @@ module application-generator
 go 1.20
 
 require (
+	github.com/iancoleman/strcase v0.3.0
 	github.com/spf13/cobra v1.7.0
 	gopkg.in/yaml.v3 v3.0.1
 	k8s.io/apimachinery v0.27.3
